Better matching with auxiliary reassociation
Allow matching things like `(+ (* a b) (* c d))` on `(+ x (* a b) (* c d))`.
Which was nearly impossible before if we take a purely decomposition approach.
What we do now is try to find subset of the root reduction elements that match
and then pull them out. So in this example, we would rewrite the root reduction as
(+ x t) where t = (+ (* a b) (* c d)).